chore(refs T30969): move DpTreeList to components root
