+++ src/components/DpTreeList/DpTreeList.vue
@@ -81,9 +81,9 @@
 </template>
 
 <script>
-import { deepMerge, hasOwnProp } from '../../../utils'
-import DpDraggable from '../../DpDraggable/DpDraggable'
-import { Stickier } from '../../../lib'
+import { deepMerge, hasOwnProp } from '../../utils'
+import DpDraggable from '../DpDraggable/DpDraggable'
+import { Stickier } from '../../lib'
 import bus from './utils/bus'
 import DpTreeListCheckbox from './DpTreeListCheckbox'
 import DpTreeListNode from './DpTreeListNode'
