+++ src/components/DpTreeList/DpTreeListNode.vue
@@ -102,11 +102,11 @@
 <script>
 import { checkboxWidth, dragHandleWidth, levelIndentationWidth } from './utils/constants'
 import bus from './utils/bus'
-import DpDraggable from '../../DpDraggable/DpDraggable'
-import DpIcon from '../../DpIcon/DpIcon'
+import DpDraggable from '../DpDraggable/DpDraggable'
+import DpIcon from '../DpIcon/DpIcon'
 import DpTreeListCheckbox from './DpTreeListCheckbox'
 import DpTreeListToggle from './DpTreeListToggle'
-import { Tooltip } from '../../../directives'
+import { Tooltip } from '../../directives'
 
 export default {
   name: 'DpTreeListNode',
